+++ Kernel/Arch/Processor.h
@@ -13,7 +13,7 @@
 #if ARCH(X86_64) || ARCH(I386)
 #    include <Kernel/Arch/x86/x86Processor.h>
 #elif ARCH(AARCH64)
-#    include <Kernel/Arch/aarch64/Processor.h>
+#    include <Kernel/Arch/aarch64/aarch64Processor.h>
 #else
 #    error "Unknown architecture"
 #endif
